requirements.txt
livekit-agents>=0.8.0.dev0
livekit-plugins-deepgram==0.6.5
livekit-plugins-openai==0.8.1
livekit-plugins-silero==0.6.4
python-dotenv==1.0.0
websockets>=12,<13